backend/otolith_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.0.1
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1